[update] change type

+++ src/main/java/com/cosmost/project/comment/view/CourseReviewView.java
@@ -1,5 +1,6 @@
 package com.cosmost.project.comment.view;
 
+import com.cosmost.project.comment.infrastructure.entity.CourseReviewEntity;
 import com.cosmost.project.comment.infrastructure.entity.CourseReviewStatus;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.Builder;
@@ -20,6 +21,6 @@ public class CourseReviewView {
     private final Long reviewerId;
     private final CourseReviewStatus courseReviewStatus;
     private final String courseReviewContent;
-    private final float rate;
+    private final Integer rate;
 
 }
